+++ frontend/src/types/DonorTypes.ts
@@ -1,0 +1,12 @@
+export type DonorResponse = {
+  id: string;
+  businessName: string;
+  email: string;
+  facebookLink?: string;
+  firstName: string;
+  instagramLink?: string;
+  lastName: string;
+  phoneNumber: string;
+  role: string;
+  userId: string;
+};
